refactor(ui): split ESP flash presenters
Keep readiness and journey builders separate so the top-level\nESP flash presenter only composes stable panel sections.\n\nRefs #2486\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
